feat: move `TsJestTransformer` to legacy mode
