+++ e2e/ast-transformers/hoist-jest/jest-isolated.config.js
@@ -13,6 +13,6 @@ module.exports = {
     react$: '<rootDir>/node_modules/react',
   },
   transform: {
-    '^.+.[tj]sx?$': '<rootDir>/../../../dist/index.js',
+    '^.+.[tj]sx?$': '<rootDir>/../../../legacy.js',
   },
 }
