+++ e2e/ast-transformers/hoist-jest/package.json
@@ -16,7 +16,7 @@
       "react$": "<rootDir>/node_modules/react"
     },
     "transform": {
-      "^.+\\.[tj]sx?$": "<rootDir>/../../../dist/index.js"
+      "^.+\\.[tj]sx?$": "<rootDir>/../../../legacy.js"
     }
   }
 }
