+++ e2e/ast-transformers/transformer-in-ts/package.json
@@ -11,7 +11,7 @@
       }
     },
     "transform": {
-      "^.+\\.[tj]sx?$": "<rootDir>/../../../dist/index.js"
+      "^.+\\.[tj]sx?$": "<rootDir>/../../../legacy.js"
     }
   }
 }
