+++ e2e/ast-transformers/transformer-options/jest-isolated.config.js
@@ -16,6 +16,6 @@ module.exports = {
     },
   },
   transform: {
-    '^.+.[tj]sx?$': '<rootDir>/../../../dist/index.js',
+    '^.+.[tj]sx?$': '<rootDir>/../../../legacy.js',
   },
 }
