+++ e2e/ast-transformers/transformer-options/package.json
@@ -20,7 +20,7 @@
       }
     },
     "transform": {
-      "^.+\\.[tj]sx?$": "<rootDir>/../../../dist/index.js"
+      "^.+\\.[tj]sx?$": "<rootDir>/../../../legacy.js"
     }
   }
 }
