+++ e2e/const-enum/jest-isolated.config.js
@@ -6,6 +6,6 @@ module.exports = {
     },
   },
   transform: {
-    '^.+.tsx?$': '<rootDir>/../../dist/index.js',
+    '^.+.tsx?$': '<rootDir>/../../legacy.js',
   },
 }
