+++ e2e/coverage-report/jest-isolated.config.js
@@ -7,6 +7,6 @@ module.exports = {
     },
   },
   transform: {
-    '^.+.tsx?$': '<rootDir>/../../dist/index.js',
+    '^.+.tsx?$': '<rootDir>/../../legacy.js',
   },
 }
