+++ e2e/coverage-report/package.json
@@ -7,7 +7,7 @@
       "!**/coverage/**"
     ],
     "transform": {
-      "^.+.tsx?$": "<rootDir>/../../dist/index.js"
+      "^.+.tsx?$": "<rootDir>/../../legacy.js"
     }
   }
 }
