+++ e2e/custom-compiler/ttypescript/package.json
@@ -10,7 +10,7 @@
       }
     },
     "transform": {
-      "^.+.tsx?$": "<rootDir>/../../../dist/index.js"
+      "^.+.tsx?$": "<rootDir>/../../../legacy.js"
     }
   }
 }
