+++ e2e/diagnostics/jest-ignored-code.config.js
@@ -8,6 +8,6 @@ module.exports = {
     },
   },
   transform: {
-    '^.+.tsx?$': '<rootDir>/../../dist/index.js',
+    '^.+.tsx?$': '<rootDir>/../../legacy.js',
   },
 }
