+++ e2e/native-esm-js/isolated/package.json
@@ -13,7 +13,7 @@
       "@jest/globals": "<rootDir>/../../../node_modules/@jest/globals"
     },
     "transform": {
-      "^.+.tsx?$": "<rootDir>/../../../dist/index.js"
+      "^.+.tsx?$": "<rootDir>/../../../legacy.js"
     }
   }
 }
