+++ e2e/native-esm-ts/package.json
@@ -7,7 +7,7 @@
       }
     },
     "transform": {
-      "^.+.tsx?$": "<rootDir>/../../dist/index.js"
+      "^.+.tsx?$": "<rootDir>/../../legacy.js"
     }
   }
 }
