+++ e2e/presets/package.json
@@ -9,7 +9,7 @@
       }
     },
     "transform": {
-      "^.+.[tj]sx?$": "<rootDir>/../../dist/index.js"
+      "^.+.[tj]sx?$": "<rootDir>/../../legacy.js"
     }
   }
 }
