+++ e2e/source-map/disabled/jest-isolated.config.js
@@ -11,6 +11,6 @@ module.exports = {
     },
   },
   transform: {
-    '^.+.tsx?$': '<rootDir>/../../../dist/index.js',
+    '^.+.tsx?$': '<rootDir>/../../../legacy.js',
   },
 }
