+++ e2e/source-map/disabled/jest.config.js
@@ -10,6 +10,6 @@ module.exports = {
     },
   },
   transform: {
-    '^.+.tsx?$': '<rootDir>/../../../dist/index.js',
+    '^.+.tsx?$': '<rootDir>/../../../legacy.js',
   },
 }
