+++ e2e/source-map/enabled/jest-isolated.config.js
@@ -8,6 +8,6 @@ module.exports = {
     },
   },
   transform: {
-    '^.+.tsx?$': '<rootDir>/../../../dist/index.js',
+    '^.+.tsx?$': '<rootDir>/../../../legacy.js',
   },
 }
