+++ e2e/source-map/enabled/jest.config.js
@@ -3,6 +3,6 @@ module.exports = {
   displayName: 'sourcemap-enabled',
   roots: ['<rootDir>', '<rootDir>/../__tests__'],
   transform: {
-    '^.+.tsx?$': '<rootDir>/../../../dist/index.js',
+    '^.+.tsx?$': '<rootDir>/../../../legacy.js',
   },
 }
