+++ e2e/test-utils/package.json
@@ -1,7 +1,7 @@
 {
   "jest": {
     "transform": {
-      "^.+.tsx?$": "<rootDir>/../../dist/index.js"
+      "^.+.tsx?$": "<rootDir>/../../legacy.js"
     }
   }
 }
