+++ e2e/transform-js/no-type-check/jest.config.js
@@ -9,6 +9,6 @@ module.exports = {
     },
   },
   transform: {
-    '^.+.[tj]sx?$': '<rootDir>/../../../dist/index.js',
+    '^.+.[tj]sx?$': '<rootDir>/../../../legacy.js',
   },
 }
