+++ e2e/transform-js/type-check/jest.config.js
@@ -8,6 +8,6 @@ module.exports = {
     },
   },
   transform: {
-    '^.+.[tj]sx?$': '<rootDir>/../../../dist/index.js',
+    '^.+.[tj]sx?$': '<rootDir>/../../../legacy.js',
   },
 }
