+++ e2e/transform-tsx/jest-isolated.config.js
@@ -7,6 +7,6 @@ module.exports = {
   },
   testEnvironment: 'jsdom',
   transform: {
-    '^.+.tsx?$': '<rootDir>/../../dist/index.js',
+    '^.+.tsx?$': '<rootDir>/../../legacy.js',
   },
 }
