+++ e2e/transform-tsx/package.json
@@ -7,7 +7,7 @@
   "jest": {
     "testEnvironment": "jsdom",
     "transform": {
-      "^.+.tsx?$": "<rootDir>/../../dist/index.js"
+      "^.+.tsx?$": "<rootDir>/../../legacy.js"
     }
   }
 }
