+++ src/legacy/index.ts
@@ -1,0 +1,7 @@
+import { TsJestTransformer } from './ts-jest-transformer'
+
+export * from './ts-jest-transformer'
+
+export default {
+  createTransformer: (): TsJestTransformer => new TsJestTransformer(),
+}
